feat(redesign): remove old pages and components, update routes and RSS feed
