+++ .superpowers/brainstorm/97207-1775391441/state/server-stopped
@@ -1,0 +1,1 @@
+{"reason":"idle timeout","timestamp":1775397801969}
