--- site/src/components/SubNav.astro
@@ -1,35 +1,0 @@
----
-type NavItem = { label: string; href: string };
-
-interface Props {
-  section: "books" | "films" | "links";
-}
-
-const { section } = Astro.props;
-
-const navMap: Record<string, NavItem[]> = {
-  books: [
-    { label: "To Read", href: "/books/toread/" },
-    { label: "Reading", href: "/books/reading/" },
-    { label: "Read", href: "/books/read/" },
-  ],
-  films: [
-    { label: "To Watch", href: "/films/towatch/" },
-    { label: "Watched", href: "/films/watched/" },
-  ],
-  links: [
-    { label: "Saved Links", href: "/links/saved/" },
-    { label: "Blogroll", href: "/links/blogroll/" },
-  ],
-};
-
-const items = navMap[section] ?? [];
----
-
-<nav class="sub-nav" aria-label="Sub Navigation">
-  <ul>
-    {items.map((item) => (
-      <li><a href={item.href}>{item.label}</a></li>
-    ))}
-  </ul>
-</nav>
