--- site/src/pages/highlight/[...page].astro
@@ -1,71 +1,0 @@
----
-import { getCollection, render } from "astro:content";
-import type { GetStaticPathsOptions } from "astro";
-import Base from "../../layouts/Base.astro";
-import Pagination from "../../components/Pagination.astro";
-import Calendar from "../../components/icons/Calendar.astro";
-
-export async function getStaticPaths({ paginate }: GetStaticPathsOptions) {
-  const highlights = await getCollection("highlight");
-  const sorted = highlights.sort(
-    (a, b) => b.data.date.getTime() - a.data.date.getTime()
-  );
-  return paginate(sorted, { pageSize: 10 });
-}
-
-const { page } = Astro.props;
-
-// Render content for each highlight on this page
-const rendered = await Promise.all(
-  page.data.map(async (highlight) => {
-    const { Content } = await render(highlight);
-    return { highlight, Content };
-  })
-);
----
-
-<Base title="Highlights" section="highlight">
-  <div class="highlight-section-main">
-    <h1 class="highlight-section-title">
-      <svg
-        height="28"
-        width="28"
-        viewBox="0 0 21 21"
-        style="color: inherit"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="m9.24264069 5.05025253v7.07106777c.17157287 3.7363257-1.24264069 5.6219437-4.24264069 5.6568543-3 .0349105-4.41421356-1.8507075-4.24264069-5.6568543v-7.07106777c-.11438191-2.82842712.82842713-4.24264068 2.82842713-4.24264068s2.94280904 1.41421356 2.82842712 4.24264068v7.07106777c.07786153 1.4142136-.39354299 2.1213204-1.41421356 2.1213204s-1.49207509-.7071068-1.41421356-2.1213204v-7.07106777"
-          fill="none"
-          stroke="#2a2e3b"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-          transform="matrix(.70710678 .70710678 -.70710678 .70710678 13.535926 1.393384)"
-        />
-      </svg>
-      Highlights
-    </h1>
-    {rendered.map(({ highlight, Content }) => {
-      const dateIso = highlight.data.date.toISOString().slice(0, 10);
-      return (
-        <article class="highlight-list-item">
-          <h3 class="highlight-list-item__title">
-            <a href={`/highlight/${highlight.id}/`}>{highlight.data.title}</a>
-          </h3>
-          <p class="highlight-list-item__date">
-            <Calendar /> {dateIso}
-          </p>
-          <div class="highlight-list-item__content">
-            <Content />
-          </div>
-        </article>
-      );
-    })}
-    <Pagination
-      currentPage={page.currentPage}
-      lastPage={page.lastPage}
-      prevUrl={page.url.prev}
-      nextUrl={page.url.next}
-    />
-  </div>
-</Base>
