+++ site/src/pages/index.xml.ts
@@ -1,41 +1,35 @@
 import rss from "@astrojs/rss";
-import { getCollection } from "astro:content";
 import type { APIContext } from "astro";
+import { getCollection } from "astro:content";
 
 export async function GET(context: APIContext) {
   const posts = await getCollection("post");
   const notes = await getCollection("note");
-  const highlights = await getCollection("highlight");
   const photos = await getCollection("photo");
 
-  const allItems = [
+  const items = [
     ...posts.map((p) => ({
       title: p.data.title,
       pubDate: p.data.date,
-      link: `/post/${p.id}/`,
+      link: `/writing/${p.id}/`,
+      description: p.body?.slice(0, 200) ?? "",
     })),
     ...notes.map((n) => ({
       title: n.data.title,
       pubDate: n.data.date,
-      link: `/note/${n.id}/`,
-    })),
-    ...highlights.map((h) => ({
-      title: h.data.title,
-      pubDate: h.data.date,
-      link: `/highlight/${h.id}/`,
+      link: `/notes/${n.id}/`,
     })),
     ...photos.map((p) => ({
       title: p.data.title,
       pubDate: p.data.date,
       link: `/photo/${p.id}/`,
-      description: "This post is a photo. Visit really.lol to view it.",
     })),
   ].sort((a, b) => b.pubDate.getTime() - a.pubDate.getTime());
 
   return rss({
     title: "really.lol",
-    description: "Jack Reid's blog",
-    site: context.site!.toString(),
-    items: allItems,
+    description: "Writing, photos, notes, and things read and watched.",
+    site: context.site!,
+    items,
   });
 }
