--- site/src/pages/note/[slug].astro
@@ -1,41 +1,0 @@
----
-import { getCollection, render } from "astro:content";
-import Base from "../../layouts/Base.astro";
-import Calendar from "../../components/icons/Calendar.astro";
-
-export async function getStaticPaths() {
-  const notes = await getCollection("note");
-  return notes.map((note) => ({
-    params: { slug: note.id },
-    props: { note },
-  }));
-}
-
-const { note } = Astro.props;
-const { Content } = await render(note);
-
-const dateIso = note.data.date.toISOString().slice(0, 10);
-const dateStr = note.data.date.toLocaleDateString("en-GB", {
-  year: "numeric",
-  month: "short",
-  day: "numeric",
-});
----
-
-<Base title={note.data.title} section="note">
-  <main class="note-main">
-    <img
-      class="note-main__image"
-      alt="Cyanotype of parts of a musket gun"
-      src="/img/layout/smithsonian-tackle.png"
-    />
-    <div class="note-main__container">
-      <div class="note-main__content">{note.data.title}</div>
-      <div class="note-main__infoline">
-        <time class="note-main__date" datetime={dateIso}>
-          <Calendar /> {dateStr}
-        </time>
-      </div>
-    </div>
-  </main>
-</Base>
