--- site/src/pages/note/index.astro
@@ -1,49 +1,0 @@
----
-import { getCollection } from "astro:content";
-import Base from "../../layouts/Base.astro";
-import Pagination from "../../components/Pagination.astro";
-
-const notes = await getCollection("note");
-const sorted = notes.sort(
-  (a, b) => b.data.date.getTime() - a.data.date.getTime()
-);
-
-const pageSize = 10;
-const totalPages = Math.ceil(sorted.length / pageSize);
-const pageData = sorted.slice(0, pageSize);
----
-
-<Base title="Notes" section="note">
-  <main class="note-section-main">
-    <header class="note-section-header">
-      <img
-        alt="A statue of an angel"
-        class="section-hero-image--note"
-        src="/img/layout/angel-small.png"
-      />
-      <h1 class="note-list-title">Notes</h1>
-    </header>
-    {pageData.map((note) => {
-      const dateIso = note.data.date.toISOString().slice(0, 10);
-      const dateStr = note.data.date.toLocaleDateString("en-GB", {
-        year: "numeric",
-        month: "short",
-        day: "numeric",
-      });
-      return (
-        <article class="note-list-item">
-          <div class="note-list__content">{note.data.title}</div>
-          <time class="note-list__date" datetime={dateIso}>
-            <a href={`/note/${note.id}/`}>{dateStr}</a>
-          </time>
-        </article>
-      );
-    })}
-    <Pagination
-      currentPage={1}
-      lastPage={totalPages}
-      prevUrl={undefined}
-      nextUrl={totalPages > 1 ? "/note/page/2/" : undefined}
-    />
-  </main>
-</Base>
