--- site/src/pages/note/page/[page].astro
@@ -1,71 +1,0 @@
----
-import { getCollection } from "astro:content";
-import type { GetStaticPathsOptions } from "astro";
-import Base from "../../../layouts/Base.astro";
-import Pagination from "../../../components/Pagination.astro";
-
-export async function getStaticPaths() {
-  const notes = await getCollection("note");
-  const sorted = notes.sort(
-    (a, b) => b.data.date.getTime() - a.data.date.getTime()
-  );
-
-  const pageSize = 10;
-  const totalPages = Math.ceil(sorted.length / pageSize);
-
-  // Generate paths for pages 2+
-  const paths = [];
-  for (let page = 2; page <= totalPages; page++) {
-    const start = (page - 1) * pageSize;
-    const data = sorted.slice(start, start + pageSize);
-    paths.push({
-      params: { page: String(page) },
-      props: {
-        pageData: data,
-        currentPage: page,
-        lastPage: totalPages,
-        prevUrl: page === 2 ? "/note/" : `/note/page/${page - 1}/`,
-        nextUrl: page < totalPages ? `/note/page/${page + 1}/` : undefined,
-      },
-    });
-  }
-  return paths;
-}
-
-const { pageData, currentPage, lastPage, prevUrl, nextUrl } = Astro.props;
----
-
-<Base title="Notes" section="note">
-  <main class="note-section-main">
-    <header class="note-section-header">
-      <img
-        alt="A statue of an angel"
-        class="section-hero-image--note"
-        src="/img/layout/angel-small.png"
-      />
-      <h1 class="note-list-title">Notes</h1>
-    </header>
-    {pageData.map((note: any) => {
-      const dateIso = note.data.date.toISOString().slice(0, 10);
-      const dateStr = note.data.date.toLocaleDateString("en-GB", {
-        year: "numeric",
-        month: "short",
-        day: "numeric",
-      });
-      return (
-        <article class="note-list-item">
-          <div class="note-list__content">{note.data.title}</div>
-          <time class="note-list__date" datetime={dateIso}>
-            <a href={`/note/${note.id}/`}>{dateStr}</a>
-          </time>
-        </article>
-      );
-    })}
-    <Pagination
-      currentPage={currentPage}
-      lastPage={lastPage}
-      prevUrl={prevUrl}
-      nextUrl={nextUrl}
-    />
-  </main>
-</Base>
