--- site/src/pages/photo/page/[page].astro
@@ -1,74 +1,0 @@
----
-import { getCollection } from "astro:content";
-import Base from "../../../layouts/Base.astro";
-import Pagination from "../../../components/Pagination.astro";
-import Calendar from "../../../components/icons/Calendar.astro";
-import Location from "../../../components/icons/Location.astro";
-import { r2Thumb } from "../../../lib/images";
-
-export async function getStaticPaths() {
-  const photos = await getCollection("photo");
-  const sorted = photos.sort(
-    (a, b) => b.data.date.getTime() - a.data.date.getTime()
-  );
-
-  const pageSize = 24;
-  const totalPages = Math.ceil(sorted.length / pageSize);
-
-  const paths = [];
-  for (let page = 2; page <= totalPages; page++) {
-    const start = (page - 1) * pageSize;
-    const data = sorted.slice(start, start + pageSize);
-    paths.push({
-      params: { page: String(page) },
-      props: {
-        pageData: data,
-        currentPage: page,
-        lastPage: totalPages,
-        prevUrl: page === 2 ? "/photo/" : `/photo/page/${page - 1}/`,
-        nextUrl: page < totalPages ? `/photo/page/${page + 1}/` : undefined,
-      },
-    });
-  }
-  return paths;
-}
-
-const { pageData, currentPage, lastPage, prevUrl, nextUrl } = Astro.props;
----
-
-<Base title="Photos" section="photo">
-  <main class="photo-section-main">
-    <h1 class="default-section-title">Photos</h1>
-    <div class="photo-section-grid">
-      {pageData.map((photo: any) => {
-        const dateIso = photo.data.date.toISOString().slice(0, 10);
-        const thumbUrl = r2Thumb(photo.data.image);
-        return (
-          <article class="photo-section-item">
-            <a href={`/photo/${photo.id}/`}>
-              <img src={thumbUrl} alt={photo.data.title} />
-            </a>
-            <h3 class="photo-section-item__title">
-              <a href={`/photo/${photo.id}/`}>{photo.data.title}</a>
-            </h3>
-            <p class="photo-section-item__date">
-              <Calendar /> {dateIso}
-            </p>
-            {photo.data.location && (
-              <p class="photo-section-item__location">
-                <Location />
-                <span>{photo.data.location}</span>
-              </p>
-            )}
-          </article>
-        );
-      })}
-    </div>
-    <Pagination
-      currentPage={currentPage}
-      lastPage={lastPage}
-      prevUrl={prevUrl}
-      nextUrl={nextUrl}
-    />
-  </main>
-</Base>
