--- site/src/pages/post/index.astro
@@ -1,45 +1,0 @@
----
-import { getCollection } from "astro:content";
-import Base from "../../layouts/Base.astro";
-import PreviewCard from "../../components/PreviewCard.astro";
-import Pagination from "../../components/Pagination.astro";
-import { generateSummary } from "../../lib/summary";
-
-const posts = await getCollection("post");
-const sorted = posts.sort(
-  (a, b) => b.data.date.getTime() - a.data.date.getTime()
-);
-
-const pageSize = 10;
-const totalPages = Math.ceil(sorted.length / pageSize);
-const pageData = sorted.slice(0, pageSize);
----
-
-<Base title="Posts" section="post">
-  <main class="section-main section-main--post">
-    <img
-      alt="A watercolor illustration of mushrooms"
-      class="section-hero-image--post"
-      src="/img/layout/mushrooms.png"
-    />
-    <h1 class="section-title">Posts</h1>
-    <div class="multi-preview-list">
-      {pageData.map((post) => (
-        <PreviewCard
-          type="post"
-          title={post.data.title}
-          date={post.data.date}
-          id={post.id}
-          tags={post.data.tags}
-          excerpt={generateSummary(post.body ?? "")}
-        />
-      ))}
-    </div>
-    <Pagination
-      currentPage={1}
-      lastPage={totalPages}
-      prevUrl={undefined}
-      nextUrl={totalPages > 1 ? "/post/page/2/" : undefined}
-    />
-  </main>
-</Base>
